+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added 
-- Testing: Python Interface for AI Models
+- Python Interface for Online AI Models (W.I.P)
 - Status & -count to Battler Log
 ### Removed
 ### Changed